refactor(domain): split crawler_entities.rs into cohesive modules
Phase 1: Domain isolation — split 750-line monolith into 5 focused modules:

  src/domain/site/config.rs          → CrawlerConfig, CrawlerConfigBuilder
  src/domain/crawl_job/entities.rs   → DiscoveredUrl, ContentType
  src/domain/result/crawl_result.rs  → CrawlResult
  src/domain/error/crawl_error.rs    → CrawlError (14 variants)
  src/domain/pattern_matching/       → matches_pattern (SSRF-safe)

  src/domain/crawler_entities.rs     → Re-export facade (backward compat)

GitNexus impact analysis:
  - matches_pattern: CRITICAL (6 impacted, 3 modules) — kept public API
  - CrawlerConfig: LOW (2 impacted, domain only)
  - CrawlError/DiscoveredUrl/CrawlResult: ZERO impact — isolated

All 448 tests pass. just check: PASS (zero errors).
No public API breakage — re-exports maintain backward compatibility.
