+++ src/domain/crawl_job/mod.rs
@@ -1,0 +1,7 @@
+//! Crawl job entities
+//!
+//! Types representing URLs and content discovered during crawling.
+
+mod entities;
+
+pub use entities::{ContentType, DiscoveredUrl};
