+++ src/domain/error/mod.rs
@@ -1,0 +1,7 @@
+//! Crawl error types
+//!
+//! Error types for the crawling domain layer.
+
+mod crawl_error;
+
+pub use crawl_error::CrawlError;
