+++ src/domain/result/mod.rs
@@ -1,0 +1,5 @@
+//! Crawl result type
+
+mod crawl_result;
+
+pub use crawl_result::CrawlResult;
