+++ src/domain/site/mod.rs
@@ -1,0 +1,7 @@
+//! Site configuration
+//!
+//! Configuration and builder for crawling a specific site.
+
+mod config;
+
+pub use config::{CrawlerConfig, CrawlerConfigBuilder};
